Move DocumentingStore into its own files
